Add all art using LFS and complete messenger script

+++ bot/messenger.py
@@ -11,11 +11,27 @@ from discord import Client, Message, User
 
 # Initialise logging and include Discord SDK logs.
 
-logger: Logger = logging.getLogger(__name__)
+LOGGER: Logger = logging.getLogger(__name__)
 
 logging.basicConfig(level=logging.INFO)
+log_formatter: logging.Formatter = logging.Formatter(
+    "(%(asctime)s) [%(levelname)s]: %(message)s",
+    "%Y-%m-%d %H:%M:%S",
+)
+
+stream_handler: logging.StreamHandler = logging.StreamHandler()
+
+stream_handler.setLevel(logging.INFO)
+stream_handler.setFormatter(log_formatter)
+
+logging.root.handlers.clear()
+
+LOGGER.addHandler(stream_handler)
 
 # Constants.
+
+CLIENT_SECRET_KEY: str = "CLIENT_SECRET"
+"""The name of the environment variable used to feed to the Discord SDK."""
 
 MESSAGE_HISTORY_UPPER_LIMIT: int = 10000
 """An upper limit of the number of messages to retrieve."""
@@ -25,6 +41,8 @@ SUBMISSIONS_CHANNEL = "submissions"
 
 
 def do_dump_all_messages():
+    """Dump all messages from all connected Discord guilds."""
+
     # Create a closure that creates a Discord SDK client.
 
     client: Client = discord.Client()
@@ -46,6 +64,8 @@ def do_dump_all_messages():
 
         # Find a channel named "submissions" in the connected guild(s):
 
+        LOGGER.info("Connecting to channel...")
+
         channel: Optional[Any] = discord.utils.get(
             client.get_all_channels(), name=SUBMISSIONS_CHANNEL
         )
@@ -59,6 +79,8 @@ def do_dump_all_messages():
             limit=MESSAGE_HISTORY_UPPER_LIMIT
         ).flatten()
 
+        LOGGER.info("Writing attachments and messages...")
+
         message: Message
         for message in all_messages:
             created_timestamp: str = message.created_at.isoformat()
@@ -67,6 +89,8 @@ def do_dump_all_messages():
             data["users"].add(author_with_discriminator)
 
             # Save all attachments to disk and then reference it.
+
+            LOGGER.debug("Writing attachments for user: [%s]...", author_with_discriminator)
 
             attachments: List[Dict[str, str]] = []
             for attachment in message.attachments:
@@ -77,13 +101,18 @@ def do_dump_all_messages():
                 if not os.path.exists(author_directory):
                     os.makedirs(author_directory)
 
+                # Save the file locally. If it already exists, skip.
+
                 filename: str = f"{created_timestamp}+{attachment.filename}"
                 local_path: str = os.path.join(author_directory, filename)
 
-                with open(local_path, "wb") as attachment_file:
-                    logger.info("Writing: [%s]...", local_path)
+                if not os.path.exists(local_path):
+                    with open(local_path, "wb") as attachment_file:
+                        LOGGER.info("Writing: [%s]...", local_path)
 
-                    await attachment.save(attachment_file)
+                        await attachment.save(attachment_file)
+                else:
+                    LOGGER.debug("Exists, skipping: [%s]...", local_path)
 
                 attachments_map: dict = {
                     "id": attachment.id,
@@ -118,20 +147,44 @@ def do_dump_all_messages():
         data["user_count"] = len(data["users"])
         data["attachment_count"] = len(data["attachments"])
 
+        LOGGER.info(
+            "Finished writing all [%d] attachments and [%d] messages!",
+            data["attachment_count"],
+            data["message_count"],
+        )
+
         # Write to file.
 
-        data["users"] = sorted(data["users"])
+        LOGGER.info("Sorting [%d] users...", data["user_count"])
+
+        data["users"] = sorted(list(data["users"]))
+
+        LOGGER.info("Writing retrieved.json file...")
 
         with open("out/retrieved.json", "w") as json_file:
             json.dump(data, json_file, indent=2)
 
+        LOGGER.info("Done with Discord, closing client!")
+
         await client.close()
 
-    client.run("ODM1MTg3MzU2NDYxODI2MDk4.YILy1g.GC32acHsgcE203sXTV1Rqkph8Xc")
+    client.run(os.environ.get(CLIENT_SECRET_KEY))
 
 
 def get_name_with_discriminator(user: User) -> str:
-    """Retrieve a name with discriminator from a user."""
+    """
+    Retrieve a name with discriminator from a user.
+
+    Parameters
+    ----------
+    user : `User`
+        The `User` whose name we want.
+
+    Returns
+    -------
+    `str`
+        A username with a discriminator separated by a "#" symbol.
+    """
 
     return f"{user.name}#{user.discriminator}"
